Createed POSIX_threads_1.text
This set of prompts make ChatGPT write a demonstration code for performing sweeps with POSIX threads, in the sense of "loop-based parallelization", similar to OpenMP. The idea was to find a way to make ChatGPT recognize that repeatedly spawning threads has an associated cost, and avoid doing this. In reality, the cost is not only associated with the actual spawning of the threads, but with the fact that in a multi-socket NUMA system the thread's stack will be allocated at the spawning caller, and therefore will have memory affinity to a single socket. Spawning threads ahead of time, and managing affinity, allows for avoiding working across NUMA nodes. Of course we would not use this to do OpenMP style work across NUMA nodes; we would use POSIX threads in independent data blocks.

What ultimately enables all of this is the barrier construct, which I asked ChatGPT to give me an explanation for. The important thing to note is that the "wait" function does _not_ poll.

IN 2023/05/08